feat: Update TV power commands and add volume commands
This commit updates the `conf/commands.yaml` file to replace the PowerOff command URI with `ssap://com.webos.service.tvpower/power/turnOffScreen` and adds a `standbyMode` payload. It also adds a PowerOn command with a similar URI and payload. In addition, it introduces VolumeDown and VolumeUp commands with the appropriate URIs. The `pkg/commands.go` file is modified to rename the `Commands` receiver in the `ReadCommands` method to `cs`. The `pkg/lsd.go` file is modified to update the log messages for creating the lsd manager and reading the conf file. The `src/main.go` file is modified to add a loop that receives messages from the websocket connection until it is closed, printing the received message. It also adds a log message to indicate when the service is shutting down.