feat: Add support for reading config path from command line
This commit adds a new flag `cfg` to the `main.go` file, allowing the user to specify the path to the config file. The flag is defined using `flag.StringVar` and the value is stored in the `cfg` variable. The `CreateLsd` function in `pkg/lsd.go` has been updated to accept the `cfg` parameter and set the `ConfigPath` field accordingly. If `cfg` is not provided, the default config path is set to the user's config directory followed by "/LGTVShutDown/config.yaml". The `LoadConfig` and `SaveConfig` functions in `pkg/config.go` now use the `ConfigPath` field instead of a hardcoded path.

The changes in this commit allow for more flexibility in specifying the path to the config file and make it easier to read and write the configuration.