Update LGTVShutDowner functionality and installation process
Description:
This commit updates the LGTVShutDowner tool to enhance its functionality. It creates a system tray icon that allows users to send different commands to their WebOS screen. The tool can now be run as a service, automatically turning the screen on at login when the `-cmd PowerOn` argument is provided. The commit also includes an updated installation process.

Changes include:
- Improved functionality of the LGTVShutDowner tool
- Added support for running the tool as a service
- Updated installation instructions and configuration files

Please note that the tool is only intended to work on Linux. Windows users are directed to the Maassoft/ColorControl repository for similar functionality.

Installation:
- Download the latest release from [here](https://github.com/Prouk/LGTVShutDowner/releases) or build the binary from source
- Follow the `.config` structure provided in the archive
- Create a LGTVShutDown folder in '/opt/' and copy the binary file into it
- Complete the config file with your LGTVWebOs IP (for normal communication) and MAC address (for wake on LAN)
- Verify and complete the LGTVShutDown.service file to match your file paths
- Register and launch the service using the following command: `systemctl --user enable LGTVShutDowner --now`